Module moore_vhdl_syntax::lexer::bundler

Structs

Bundler
A grinder that bundles up categorized characters into groups and converts spaces and comments into single tokens, dropping any information about their content. String and bit string literals are formed here as well.

Enums

Bundle
A bundle of characters. These are the most fundamental groups of characters as per the VHDL standard. Lexical analysis will aggregate one or more of these into more meaningful tokens.